requirements.txt
aiogram==3.0.0b7
colorama==0.4.6
python-dotenv==1.0.0
aiosqlite==0.19.0
selenium==4.12.0
